[GBA-9] Allow multiple plans
Schedule plan runs with delay for command and comment executions
