+++ lib/github_ci_app.rb
@@ -45,6 +45,8 @@ require_relative '../workers/timeout_execution'
 require_relative '../workers/ci_job_fetch_topotest_failures'
 require_relative '../workers/slack_username2_id'
 require_relative '../workers/create_execution_by_plan'
+require_relative '../workers/create_execution_by_comment'
+require_relative '../workers/create_execution_by_command'
 
 # Slack libs
 require_relative 'slack/slack'
